% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAYESDEF.R
\name{BAYESDEF}
\alias{BAYESDEF}
\title{Bayesian Analysis of DSD}
\usage{
BAYESDEF()
}
\value{
BAYESDEF is a graphic interface
}
\description{
Definitive Screening Designs are a class of experimental designs that under factor sparsity have the potential to estimate linear, quadratic and interaction effects with little experimental effort. BAYESDEF is a package that performs a five step strategy to analyze this kind of experiments that makes use of tools coming from the Bayesian approach. It also includes the least absolute shrinkage and selection operator (lasso) as a check (Aguirre VM. (2016) <DOI:10.1002/asmb.2160>).

You can learn more about this package at:
http://www.uv.mx/personal/nehuerta/bayesdef/
}
\details{
BAYESDEF is a package with a graphical interface dedicated to perform Bayesian analysis of
Definitive Screening Designs with thirteen runs. These very economic experimental plans are
gaining popularity because, under certain conditions, they allow the estimation of main,
interaction and quadratic effects. Tinhe package also allows the user to fit custom models to
the data. It also includes the additional feature to analyze the data using the least absolute
shrinkage and selection operator "lasso".
Note: BAYESDEF is free software and comes with ABSOLUTELY NO WARRANTY.
}
\examples{
\dontrun{
##Install package
library(BAYESDEF)
##Call the package
BAYESDEF()
}

}
\references{
Aguirre VM. Bayesian analysis of definitive screening designs when the response
is nonnormal. Applied Stochastic Models in Business and Industry 2016; 32(4):440–452.
DOI: 10.1002/asmb.2160

Aguirre VM, de la Vara R. A Bayesian analysis of very small unreplicated experiments.
Quality and Reliability Engineering International 2014a; 30(3):413–426.
DOI: 10.1002/qre.1578

Friedman J, Hastie T, Tibshirani R. Regularization paths for generalized linear
models via coordinate descent. Journal of Statistical Software 2010; 33:1–22.
DOI: 10.18637/jss.v033.i01

Jones B, Nachtsheim C. A class of three-level designs for definitive screening in
the presence of second order effects. Journal of Quality Technology 2011; 43:1–15.

Tibshirani R. Regression shrinkage and selection via the lasso. Journal of the
Royal Statistical Society B 1996; 58:267–288. DOI:10.1111/j.1467-9868.2011.00771.x
}
\author{
Victor Manuel Aguirre-Torres, Nery Sofia Huerta-Pacheco, Edgar A. Lopez
}
